add simple Qt-based model/view demo
Compared to defining a change notification Qt signal in the model and connecting
that to the Label, we can see that with qruby-cells the code becomes
less cluttered. We simply state the logical relationship between model
and view, i.e. that output.text should always be a certain function of
the model state, instead of having to figure out how to keep the output
in sync with the model.

While in this simple case only a couple of lines have been saved,
imagine what happens if your view has multiple output elements depending
on multiple model properties each.