Use weak references to track observers.
When observers or objects with dependent cells were no longer needed,
they could not be disposed of due to their observees keeping a reference
to them around. For observers, it was possible to "unobserve" cells,
which depends on the caller remembering to do this, producing hard to
detect memory leaks. For objects with dependent cells, there was no way
to get them garbage collected while one of their observees was still
alive.

Solution: Use WeakRef to reference observer blocks and objects with
dependent cells. Make Object.observe return the block, so that the
caller can easily keep a reference to it as long as needed (which would
be necessary also for explicit de-registration).